Option to disable mousewheel
I think you should provide this option.  changing values with the
mousewheel is maybe nice, nevertheless it can cause uncommon behavior:
- your mouse is somewhere on the screen
- you scroll down or up by using the mouswheel
- now it can happen that the mousepointer unfortunately stays over an
touchspin input
==> it stops scrolling the screen and changing the value instead. and
maybe the user did not realize this because he moves the mouse a bit
while scrolling (or because scrolling has stopped) and the value is
accidentally changed.